refactor: move store migrations into focused file
